+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Use GitHub as source server (#12)
 
 ### Changed
+- Rename DegreeKelvin to Kelvin and delete "degK" unit (#30)
 - Case sensitive unit prefixes
 - Switch PCL project to profile 328
 - Refactor UnitProvider.RegisterUnits to extension methods (#6)